+++ extractedTranslations/en/profile.json
@@ -1,33 +1,34 @@
 {
-    "URL Аватара": "",
-    "Аватар": "",
-    "Валюта": "",
-    "Возраст": "",
-    "Вы не можете оценить свой профиль": "Вы не можете оценить свой профиль",
-    "Город": "",
-    "Данные не указаны": "",
-    "Имя": "",
-    "Неккоректное имя или фамилия": "",
-    "Неккоректный возраст": "",
-    "Неккоректный город": "",
-    "Неккоректный псевдоним": "",
-    "Некоректное имя или фамилия": "Некоректное имя или фамилия",
-    "Некоректный возраст": "Некоректный возраст",
-    "Некоректный город": "Некоректный город",
-    "Некоректный псевдоним": "Некоректный псевдоним",
-    "Оставьте свой отзыв о профиле": "Оставьте свой отзыв о профиле",
-    "Отменить": "",
-    "Оцените профиль": "Оцените профиль",
-    "Ошибка загрузки рейтинга": "",
-    "Ошибка при получении профиля": "",
-    "Попробуйте обновить страницу": "",
-    "Произошла ошибка при загрузке профиля": "",
-    "Профиль": "",
-    "Псевдоним": "",
-    "Редактировать": "",
-    "Серверная ошибка": "",
-    "Сохранить": "",
-    "Ссылка на аватар": "",
-    "Страна": "",
-    "Фамилия": ""
+  "Fватар": "Fватар",
+  "URL Аватара": "",
+  "Аватар": "",
+  "Валюта": "",
+  "Возраст": "",
+  "Вы не можете оценить свой профиль": "Вы не можете оценить свой профиль",
+  "Город": "",
+  "Данные не указаны": "",
+  "Имя": "",
+  "Неккоректное имя или фамилия": "",
+  "Неккоректный возраст": "",
+  "Неккоректный город": "",
+  "Неккоректный псевдоним": "",
+  "Некоректное имя или фамилия": "Некоректное имя или фамилия",
+  "Некоректный возраст": "Некоректный возраст",
+  "Некоректный город": "Некоректный город",
+  "Некоректный псевдоним": "Некоректный псевдоним",
+  "Оставьте свой отзыв о профиле": "Оставьте свой отзыв о профиле",
+  "Отменить": "",
+  "Оцените профиль": "Оцените профиль",
+  "Ошибка загрузки рейтинга": "",
+  "Ошибка при получении профиля": "",
+  "Попробуйте обновить страницу": "",
+  "Произошла ошибка при загрузке профиля": "",
+  "Профиль": "",
+  "Псевдоним": "",
+  "Редактировать": "",
+  "Серверная ошибка": "",
+  "Сохранить": "",
+  "Ссылка на аватар": "",
+  "Страна": "",
+  "Фамилия": ""
 }
