+++ extractedTranslations/en/translation.json
@@ -28,7 +28,9 @@
     "Ошибка загрузки оповещений": "Ошибка загрузки оповещений",
     "Ошибка при загрузке статей": "Ошибка при загрузке статей",
     "Поиск": "Поиск",
+    "Попробуйте обновить страницу": "Попробуйте обновить страницу",
     "Произошла непредвиденная ошибка": "Произошла непредвиденная ошибка",
+    "Произошла ошибка при загрузке профиля": "Произошла ошибка при загрузке профиля",
     "Профиль": "Профиль",
     "Редактирование": "Редактирование",
     "Редактирование статьи с ID": "Редактирование статьи с ID",
