+++ extractedTranslations/ru/profile.json
@@ -1,27 +1,29 @@
 {
-    "Валюта": "Валюта",
-    "Возраст": "Возраст",
-    "Вы не можете оценить свой профиль": "Вы не можете оценить свой профиль",
-    "Город": "Город",
-    "Данные не указаны": "Данные не указаны",
-    "Имя": "Имя",
-    "Некоректное имя или фамилия": "Некоректное имя или фамилия",
-    "Некоректный возраст": "Некоректный возраст",
-    "Некоректный город": "Некоректный город",
-    "Некоректный псевдоним": "Некоректный псевдоним",
-    "Оставьте свой отзыв о профиле": "Оставьте свой отзыв о профиле",
-    "Отменить": "Отменить",
-    "Оцените профиль": "Оцените профиль",
-    "Ошибка загрузки рейтинга": "Ошибка загрузки рейтинга",
-    "Ошибка при получении профиля": "Ошибка при получении профиля",
-    "Попробуйте обновить страницу": "Попробуйте обновить страницу",
-    "Произошла ошибка при загрузке профиля": "Произошла ошибка при загрузке профиля",
-    "Профиль": "Профиль",
-    "Псевдоним": "Псевдоним",
-    "Редактировать": "Редактировать",
-    "Серверная ошибка": "Серверная ошибка",
-    "Сохранить": "Сохранить",
-    "Ссылка на аватар": "Ссылка на аватар",
-    "Страна": "Страна",
-    "Фамилия": "Фамилия"
+  "Fватар": "Fватар",
+  "Аватар": "Аватар",
+  "Валюта": "Валюта",
+  "Возраст": "Возраст",
+  "Вы не можете оценить свой профиль": "Вы не можете оценить свой профиль",
+  "Город": "Город",
+  "Данные не указаны": "Данные не указаны",
+  "Имя": "Имя",
+  "Некоректное имя или фамилия": "Некоректное имя или фамилия",
+  "Некоректный возраст": "Некоректный возраст",
+  "Некоректный город": "Некоректный город",
+  "Некоректный псевдоним": "Некоректный псевдоним",
+  "Оставьте свой отзыв о профиле": "Оставьте свой отзыв о профиле",
+  "Отменить": "Отменить",
+  "Оцените профиль": "Оцените профиль",
+  "Ошибка загрузки рейтинга": "Ошибка загрузки рейтинга",
+  "Ошибка при получении профиля": "Ошибка при получении профиля",
+  "Попробуйте обновить страницу": "Попробуйте обновить страницу",
+  "Произошла ошибка при загрузке профиля": "Произошла ошибка при загрузке профиля",
+  "Профиль": "Профиль",
+  "Псевдоним": "Псевдоним",
+  "Редактировать": "Редактировать",
+  "Серверная ошибка": "Серверная ошибка",
+  "Сохранить": "Сохранить",
+  "Ссылка на аватар": "Ссылка на аватар",
+  "Страна": "Страна",
+  "Фамилия": "Фамилия"
 }
