+++ public/locales/en/profile.json
@@ -18,7 +18,6 @@
     "Редактировать": "Edit",
     "Серверная ошибка": "Server error",
     "Сохранить": "Save",
-    "Ссылка на аватар": "Avatar Link",
     "Страна": "Country",
     "Фамилия": "Last name",
     "Вы не можете оценить свой профиль": "You can't rate your profile",
