+++ public/locales/ru/profile.json
@@ -18,7 +18,6 @@
     "Редактировать": "Редактировать",
     "Серверная ошибка": "Серверная ошибка",
     "Сохранить": "Сохранить",
-    "Ссылка на аватар": "Ссылка на аватар",
     "Страна": "Страна",
     "Фамилия": "Фамилия",
     "Вы не можете оценить свой профиль": "Вы не можете оценить свой профиль",
