+++ src/entities/Profile/module/types/profileProps.ts
@@ -1,0 +1,20 @@
+import { Country } from '@/entities/Country';
+import { Currency } from '@/entities/Currency';
+
+import { Profile } from './profile';
+
+export interface ProfileCardProps {
+    className?: string;
+    data?: Profile;
+    error?: string;
+    isLoading?: boolean;
+    readonly?: boolean;
+    onChangeUsername?: (value?: string) => void;
+    onChangeAvatar?: (value?: string) => void;
+    onChangeFirstname?: (value?: string) => void;
+    onChangeLastname?: (value?: string) => void;
+    onChangeCity?: (value?: string) => void;
+    onChangeAge?: (value?: string) => void;
+    onChangeCurrency?: (currency?: Currency) => void;
+    onChangeCountry?: (country?: Country) => void;
+}
