+++ src/entities/Profile/ui/ProfileCardDeprecated/ProfileCardDeprecated.module.scss
@@ -15,9 +15,3 @@
 .input {
     min-width: 100%;
 }
-
-.ProfileCardRedesigned {
-    .input {
-        min-width: 0;
-    }
-}
