+++ src/entities/Profile/ui/ProfileCardDeprecated/ProfileCardDeprecated.tsx
@@ -1,0 +1,132 @@
+import { useTranslation } from 'react-i18next';
+
+import { CountrySelect } from '@/entities/Country';
+import { CurrencySelect } from '@/entities/Currency';
+
+import { Mods, classNames } from '@/shared/libs/classNames/classNames';
+import { Avatar } from '@/shared/ui/deprecated/Avatar';
+import { Input } from '@/shared/ui/deprecated/Input';
+import { Loader } from '@/shared/ui/deprecated/Loader';
+import { Text, TextAlign, TextTheme } from '@/shared/ui/deprecated/Text';
+import { HStack, VStack } from '@/shared/ui/redesigned/Stack';
+
+import { ProfileCardProps } from '../../module/types/profileProps';
+import cls from './ProfileCardDeprecated.module.scss';
+
+export const ProfileCardDeprecatedError = () => {
+    const { t } = useTranslation('profile');
+    return (
+        <HStack
+            justify="center"
+            max
+            className={classNames(cls.ProfileCard, {}, [cls.error])}
+        >
+            <Text
+                theme={TextTheme.ERROR}
+                title={t('Произошла ошибка при загрузке профиля')}
+                text={t('Попробуйте обновить страницу')}
+                align={TextAlign.CENTER}
+            />
+        </HStack>
+    );
+};
+
+export const ProfileCardDeprecatedLoading = () => (
+    <HStack
+        justify="center"
+        max
+        className={classNames(cls.ProfileCard, { [cls.loading]: true }, [])}
+    >
+        <Loader />
+    </HStack>
+);
+
+export const ProfileCardDeprecated = (props: ProfileCardProps) => {
+    const {
+        className,
+        data,
+        readonly,
+        onChangeUsername,
+        onChangeAvatar,
+        onChangeFirstname,
+        onChangeLastname,
+        onChangeCity,
+        onChangeAge,
+        onChangeCurrency,
+        onChangeCountry,
+    } = props;
+
+    const { t } = useTranslation('profile');
+
+    const mods: Mods = {
+        [cls.editing]: !readonly,
+    };
+
+    return (
+        <VStack
+            max
+            gap="16"
+            className={classNames(cls.ProfileCard, mods, [className])}
+        >
+            {data?.avatar && (
+                <HStack justify="center" max className={cls.avatarWrapper}>
+                    <Avatar src={data.avatar} alt="avatar" size={150} />
+                </HStack>
+            )}
+            <Input
+                value={data?.username}
+                placeholder={t('Псевдоним')}
+                className={cls.input}
+                onChange={onChangeUsername}
+                readonly={readonly}
+            />
+            <Input
+                value={data?.firstname}
+                placeholder={t('Имя')}
+                className={cls.input}
+                onChange={onChangeFirstname}
+                readonly={readonly}
+                data-testid="ProfileCard.firstname"
+            />
+            <Input
+                value={data?.lastname}
+                placeholder={t('Фамилия')}
+                className={cls.input}
+                onChange={onChangeLastname}
+                readonly={readonly}
+                data-testid="ProfileCard.lastname"
+            />
+            <Input
+                value={data?.age}
+                placeholder={t('Возраст')}
+                className={cls.input}
+                onChange={onChangeAge}
+                readonly={readonly}
+            />
+            <CurrencySelect
+                value={data?.currency}
+                onChange={onChangeCurrency}
+                readonly={readonly}
+            />
+            <CountrySelect
+                value={data?.country}
+                onChange={onChangeCountry}
+                readonly={readonly}
+            />
+            <Input
+                value={data?.city}
+                placeholder={t('Город')}
+                className={cls.input}
+                onChange={onChangeCity}
+                readonly={readonly}
+            />
+            <Input
+                value={data?.avatar}
+                placeholder={t('Ссылка на аватар')}
+                className={cls.input}
+                onChange={onChangeAvatar}
+                readonly={readonly}
+            />
+        </VStack>
+    );
+};
