+++ src/entities/Profile/ui/ProfileCardRedesigned/ProfileCardRedesigned.module.scss
@@ -1,0 +1,5 @@
+.ProfileCardRedesigned {
+    .input {
+        min-width: 0;
+    }
+}
