+++ src/shared/ui/redesigned/Input/Input.module.scss
@@ -10,6 +10,18 @@
     width: 100%;
 }
 
+.s {
+    height: 32px;
+}
+
+.m {
+    height: 38px;
+}
+
+.l {
+    height: 44px;
+}
+
 .withAddonLeft {
     padding-left: 16px;
 }
