+++ src/shared/ui/redesigned/Stack/Flex/Flex.module.scss
@@ -54,6 +54,10 @@
     gap: 16px;
 }
 
+.gap24 {
+    gap: 24px;
+}
+
 .gap32 {
     gap: 32px;
 }
