+++ src/widgets/ArticlesFilters/ui/ArticlesFilters.tsx
@@ -51,6 +51,7 @@ export const ArticlesFilters = (props: ArticlesFiltersProps) => {
                 <Input
                     onChange={onChangeSearch}
                     value={search}
+                    size="s"
                     placeholder={t('Поиск')}
                     addonLeft={<Icon Svg={LoopIcon} width={14} height={14} />}
                     data-testid="ArtilesFilter.search"
